feat: add header and footer to layout
