+++ src/components/Layout/Layout.css
@@ -1,5 +1,5 @@
 .layout {
-  min-height: 80vh;
+  min-height: 90vh;
   padding: 0 20px;
 }
