+++ src/components/Layout/Layout.tsx
@@ -1,11 +1,31 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import ILayout from 'interfaces/ILayout';
 import './Layout.css';
+import Header from 'components/Header';
+import Footer from 'components/Footer';
 
 export default function Layout({ children }: ILayout) {
+  const [sticky, setSticky] = useState(false);
+  const handleScroll = () => {
+    if (window.pageYOffset > 0) {
+      setSticky(true);
+    } else {
+      setSticky(false);
+    }
+  };
+
+  useEffect(() => {
+    window.addEventListener('scroll', handleScroll);
+    return () => {
+      window.removeEventListener('scroll', handleScroll);
+    };
+  });
+
   return (
     <>
-      <main>{children}</main>
+      <Header isSticky={sticky} />
+      <main className="layout">{children}</main>
+      <Footer />
     </>
   );
 }
